getLong should return a long instead of int
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@192 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/include/tntdb/decimal.h
@@ -183,7 +183,7 @@ namespace tntdb
     /// Return this decimal number rounded as a C++ long.
     /// @return long result if the result will fit.
     /// @throw std::overflow_error if the result will not fit
-    int getLong() const throw(std::overflow_error)
+    long getLong() const throw(std::overflow_error)
       { return getInteger<long>(); }
 
     /// Return this decimal number rounded as a C++ int32_t.
